node files rearranged
